move sync stuff to new syncController
